First draft of calling generate from KILN
