+++ package.json
@@ -23,7 +23,8 @@
     "react": "^18.3.1",
     "react-currency-mask": "^1.3.2",
     "react-dom": "^18.3.1",
-    "react-input-mask": "^3.0.0-alpha.2"
+    "react-input-mask": "^3.0.0-alpha.2",
+    "react-router-dom": "^6.26.2"
   },
   "devDependencies": {
     "@playwright/test": "^1.46.1",
