+++ src/EditFormPage.tsx
@@ -1,0 +1,12 @@
+import "./App.css";
+import "@carbon/styles/css/styles.css";
+
+
+const EditFormPage: React.FC = () => {
+  return(
+
+    <div>Edit Page...</div>
+  );
+};
+
+export default EditFormPage;
